Simplify testing setup
