+++ cli/common.go
@@ -13,10 +13,10 @@ import (
 	"github.com/urfave/cli"
 )
 
-// remoteClient returns a remote client that connects to the service at the
+// makeRemoteClient returns a remote client that connects to the service at the
 // given host and port, with the optional SSL/TLS credentials file specified at
 // the given location.
-func remoteClient(ctx context.Context, service, host string, port int, certFilePath string) (jasper.RemoteClient, error) {
+func makeRemoteClient(ctx context.Context, service, host string, port int, certFilePath string) (jasper.RemoteClient, error) {
 	addr, err := net.ResolveTCPAddr("tcp", fmt.Sprintf("%s:%d", host, port))
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to resolve address")
@@ -68,7 +68,7 @@ func withConnection(ctx context.Context, c *cli.Context, operation func(jasper.R
 	service := c.String(serviceFlagName)
 	certFilePath := c.String(certFilePathFlagName)
 
-	client, err := remoteClient(ctx, service, host, port, certFilePath)
+	client, err := makeRemoteClient(ctx, service, host, port, certFilePath)
 	if err != nil {
 		return errors.Wrap(err, "error setting up remote client")
 	}
